This PR includes the following changes
1) allow users to configure service endpoints (bing speech/CRIS), recognition mode, language, and output formats via USPLib.
2) Use 64-bit for time and duration offset in speech response
3) re-organize error code returned by USPlib
4) adapt uspclientconsole.exe to configure different options.

There are some work to do to fully enable different recognition mode and output format:
- support new messages in dictation/conversation mode
- retrieve detailed output format.

Related work items: #1075720

+++ source/core/include/ispxinterfaces.h
@@ -335,18 +335,18 @@ public:
     using AdditionalMessagePayload_Type = void*;
     using ErrorPayload_Type = SPXHR;
     
-    virtual void SpeechStartDetected(ISpxRecoEngineAdapter* adapter, uint32_t offset) = 0;
-    virtual void SpeechEndDetected(ISpxRecoEngineAdapter* adapter, uint32_t offset) = 0;
+    virtual void SpeechStartDetected(ISpxRecoEngineAdapter* adapter, uint64_t offset) = 0;
+    virtual void SpeechEndDetected(ISpxRecoEngineAdapter* adapter, uint64_t offset) = 0;
 
-    virtual void SoundStartDetected(ISpxRecoEngineAdapter* adapter, uint32_t offset) = 0;
-    virtual void SoundEndDetected(ISpxRecoEngineAdapter* adapter, uint32_t offset) = 0;
+    virtual void SoundStartDetected(ISpxRecoEngineAdapter* adapter, uint64_t offset) = 0;
+    virtual void SoundEndDetected(ISpxRecoEngineAdapter* adapter, uint64_t offset) = 0;
 
-    virtual void IntermediateResult(ISpxRecoEngineAdapter* adapter, uint32_t offset, ResultPayload_Type payload) = 0;
-    virtual void FinalResult(ISpxRecoEngineAdapter* adapter, uint32_t offset, ResultPayload_Type payload) = 0;
+    virtual void IntermediateResult(ISpxRecoEngineAdapter* adapter, uint64_t offset, ResultPayload_Type payload) = 0;
+    virtual void FinalResult(ISpxRecoEngineAdapter* adapter, uint64_t offset, ResultPayload_Type payload) = 0;
 
     virtual void DoneProcessingAudio(ISpxRecoEngineAdapter* adapter) = 0;
 
-    virtual void AdditionalMessage(ISpxRecoEngineAdapter* adapter, uint32_t offset, AdditionalMessagePayload_Type payload) = 0;
+    virtual void AdditionalMessage(ISpxRecoEngineAdapter* adapter, uint64_t offset, AdditionalMessagePayload_Type payload) = 0;
 
     virtual void Error(ISpxRecoEngineAdapter* adapter, ErrorPayload_Type payload) = 0;
 };
